rename: `clangFormatGitPath.js` to `gitClangFormatPath.js`
