+++ packages/clang-format-git-python/src/index.js
@@ -1,4 +1,4 @@
-const { clangFormatGitPath, gitClangFormatPath } = require('./utils/clangFormatGitPath');
+const { clangFormatGitPath, gitClangFormatPath } = require('./utils/gitClangFormatPath');
 
 module.exports = {
   clangFormatGitPath,
