+++ packages/clang-format-git-python/src/utils/gitClangFormatPath.js
